Rename lib folder to src
