+++ packages/docs/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tsargp/docs",
-  "version": "1.6.1",
+  "version": "1.6.2",
   "description": "Documentation for tsargp",
   "homepage": "https://dsogari.github.io/tsargp/docs",
   "license": "MIT",
