+++ packages/docs/typedoc.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://typedoc.org/schema.json",
-  "entryPoints": ["../tsargp/lib/index.ts"],
-  "tsconfig": "../tsargp/lib/tsconfig.json",
+  "entryPoints": ["../tsargp/src/index.ts"],
+  "tsconfig": "../tsargp/src/tsconfig.json",
   "favicon": "favicon.svg",
   "out": "public/api",
   "githubPages": false,
