+++ packages/tsargp/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tsargp",
-  "version": "1.6.1",
+  "version": "1.6.2",
   "author": {
     "name": "Diego Sogari",
     "email": "diego.sogari@gmail.com",
@@ -42,7 +42,9 @@
   },
   "description": "Argument parser for TypeScript",
   "files": [
-    "dist/**/*.{js,ts}"
+    "dist/**/*.{js,ts,map}",
+    "examples",
+    "src"
   ],
   "homepage": "https://dsogari.github.io/tsargp",
   "keywords": [
