+++ packages/tsargp/test/formatter/formatter.default.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'bun:test';
-import type { Options } from '../../lib/options';
-import { format } from '../../lib/formatter';
+import type { Options } from '../../src/options';
+import { format } from '../../src/formatter';
 
 describe('format', () => {
   describe('a default value is specified', () => {
