+++ packages/tsargp/test/parser/options/parser.flag.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'bun:test';
-import { type Options } from '../../../lib/options';
-import { parse } from '../../../lib/parser';
+import { type Options } from '../../../src/options';
+import { parse } from '../../../src/parser';
 
 process.env['FORCE_WIDTH'] = '0'; // omit styles
