+++ packages/tsargp/test/parser/parser.completion.spec.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it, jest } from 'bun:test';
-import { JsonMessage, TextMessage } from '../../lib/styles';
-import { type Options } from '../../lib/options';
-import { type ParsingFlags, parse } from '../../lib/parser';
+import { JsonMessage, TextMessage } from '../../src/styles';
+import { type Options } from '../../src/options';
+import { type ParsingFlags, parse } from '../../src/parser';
 
 process.env['FORCE_WIDTH'] = '0'; // omit styles
