+++ packages/tsargp/test/parser/parser.default.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, jest } from 'bun:test';
-import { type Options, OptionValues } from '../../lib/options';
-import { parse } from '../../lib/parser';
+import { type Options, OptionValues } from '../../src/options';
+import { parse } from '../../src/parser';
 
 process.env['FORCE_WIDTH'] = '0'; // omit styles
